requirements.txt
python-dotenv
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1